New Release
This commit it's including a lot of refactoring on Lib and Interface.
It's also including a new function as Bazaar Monitoring, but the filters are not finished yet. The alert via email and sound for alert are the only concluded stuffs. The tab order was reviewed in Bazaar Form, and will be reviewed when new filters been inputed on form.